# src/rest/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: TMA App REST API Server
# during dev, should point to your local machine
host: localhost:4000
# basePath prefixes all resource paths 
basePath: /api
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  /login:
    post:
      tags:
      - "Authentication"
      description: Login using username/password
      # used as the method name of the controller
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "loginDetails"
          description: "JSON includes Username and password"
          required: true
          schema:
            type: "object"
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "404":
          description: Can not authenticate the user and password combination
          schema:
             type: "object"
        default:
          description: Error
          schema:
            type: "object"
  
  /logout:
    post:
      tags:
      - "Authentication"
      description: Log out
      # used as the method name of the controller
      consumes:
        - "application/json"
      produces:
        - "application/json"
      security:
         - Bearer: []
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "404":
          description: Can not log out
          schema:
             type: "object"
        default:
          description: Error
          schema:
            type: "object"
  
  /contact:
    get:
      tags:
      - "Contact"
      description: Get the contacts
      produces:
        - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "userId"
          description: "User Id"
          required: true
          type: "string"
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "401":
          description: Current login session is invalid
          schema:
            type: "object"
        default:
          description: Error
          schema:
            type: "object"
    post:
      tags:
      - "Contact"
      description: Create the contact
      produces:
      - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "contact"
          description: "contact"
          required: true
          schema:
            type: "object"
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "401":
          description: Current login session is invalid
          schema:
            type: "object"
        default:
          description: Error
          schema:
            type: "object"
  /contact/{contactId}:
    delete:
      tags:
      - "Contact"
      description: Get the contacts
      produces:
        - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "contactId"
          description: "Contact Id"
          required: true
          type: "string"
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "401":
          description: Current login session is invalid
          schema:
            type: "object"
        default:
          description: Error
          schema:
            type: "object"
    put:
      tags:
      - "Contact"
      description: Update the contact
      produces:
      - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "contactId"
          description: "Contact Id"
          required: true
          type: "string"
        - in: "body"
          name: "contact"
          description: "contact"
          required: true
          schema:
            type: "object"
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "401":
          description: Current login session is invalid
          schema:
            type: "object"
        default:
          description: Error
          schema:
            type: "object"            

  /message:
    get:
      tags:
      - "Message"
      description: Get the messages
      produces:
        - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "userId"
          description: "User Id"
          required: true
          type: "string"
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "401":
          description: Current login session is invalid
          schema:
            type: "object"
        default:
          description: Error
          schema:
            type: "object"
    post:
      tags:
      - "Contact"
      description: Create the message
      produces:
        - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "message"
          description: "message"
          required: true
          schema:
            type: "object"
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "401":
          description: Current login session is invalid
          schema:
            type: "object"
        default:
          description: Error
          schema:
            type: "object"
  /message/{messageId}:
    delete:
      tags:
      - "Message"
      description: Delete the message
      produces:
        - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "messageId"
          description: "message Id"
          required: true
          type: "string"
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "401":
          description: Current login session is invalid
          schema:
            type: "object"
        default:
          description: Error
          schema:
            type: "object"
            
    put:
      tags:
      - "Message"
      description: Update the message
      produces:
      - "application/json"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "messageId"
          description: "message Id"
          required: true
          type: "string"
        - in: "body"
          name: "message"
          description: "message"
          required: true
          schema:
            type: "object"
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "401":
          description: Current login session is invalid
          schema:
            type: "object"
        default:
          description: Error
          schema:
            type: "object"            
  /mailbox:
    get:
      tags:
      - "Mailbox"
      description: Get the mailboxes
      produces:
        - "application/json"
      security:
        - Bearer: []
      responses:
        "200":
          description: Success
          schema:
            type: "object"
        "401":
          description: Current login session is invalid
          schema:
            type: "object"
        default:
          description: Error
          schema:
            type: "object"